requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
